feat: implement server
closes: #5
