+++ Cargo.toml
@@ -29,5 +29,15 @@ serde = "1.0.228"
 serde_json = "1.0.149"
 smallvec = "2.0.0-alpha.12"
 tracing = "0.1.44"
+tracing-subscriber = { version = "0.3.23", default-features = false, features = [
+    "registry",
+    "fmt",
+    "std",
+    "tracing-log",
+] }
 tree-sitter = "0.26.8"
 tree-sitter-achitekfile = "0.1.0"
+
+[dev-dependencies]
+indoc = "2.0.7"
+
