+++ README.md
@@ -1,3 +1,14 @@
-# Achitekfile Language Server
+# Achitek-ls
 
+`achitek-ls` is a language server for Achitekfiles. It provides editor support
+for authoring Achitek template blueprints, including diagnostics, completion,
+hover, symbols, navigation, formatting, folding, selection ranges, and rename.
 
+> [!NOTE]
+> For supported editor features, see [CAPABILITIES.md](docs/CAPABILITIES.md).
+
+## Lend a helping hand 🫴
+
+If you'd like to contribute to `achitek-ls`, check out
+[CONTRIBUTING.md](docs/CONTRIBUTING.md). If you're simply curious about the structure
+of this project, take a peek at [ARCHITECTURE.md](docs/ARCHITECTURE.md).
