--- analysis/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "analysis"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-syntax = { path = "../syntax" }
-tree-sitter.workspace = true
