--- server/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "server"
-version = "0.1.0"
-edition = "2024"
-
-[[bin]]
-name = "achitek-ls"
-path = "src/main.rs"
-
-[dependencies]
-analysis = { path = "../analysis" }
-anyhow.workspace = true
-lsp-server.workspace = true
-lsp-types.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-syntax = { path = "../syntax" }
