+++ src/server/handlers/mod.rs
@@ -1,0 +1,15 @@
+pub mod completion;
+pub mod definition;
+pub mod diagnostics;
+pub mod did_change;
+pub mod did_close;
+pub mod did_open;
+pub mod document_symbol;
+pub mod folding_range;
+pub mod formatting;
+pub mod hover;
+pub mod prepare_rename;
+pub mod references;
+pub mod rename;
+pub mod selection_range;
+pub mod workspace_symbol;
