--- syntax/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "syntax"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-thiserror.workspace = true
-tree-sitter.workspace = true
-tree-sitter-achitekfile.workspace = true
